Moved listener config tests to config package
